refactor: shrink probes/ to minimal repeatable gate (4 files)
Removed:
- probe_cop_v2.py — merged into probe_cop.py (now canonical)
- probe_ofx.py — OFX product-declined, not a repeatable gate
- debug_cop_*.py, diagnose_ocl*.py — one-off exploration artifacts
- run_probe_v3.py — extra variant
- h21_probe_results_final.json — static artifact, no test consumer

Kept:
- probe_cop.py — canonical smoke probe (H21-corrected)
- probe_opencl_devices.py — OpenCL device enumeration
- matrix.md — compatibility matrix (H21 Windows filled)
- README.md — developer runbook with gateway path and JSON contract

Conclusion: Copernicus COP + MaterialX -> ready. OpenCL COP cook fail
on CPU-only platform -> PIP-2933 gate not yet passed.
