--- probes/h21_probe_results_final.json
@@ -1,85 +1,0 @@
-{
-  "results": [
-    {
-      "section": "host_info",
-      "status": "pass",
-      "evidence": {
-        "platform": "win32",
-        "python_version": "3.11.7 (main, Jul 30 2025, 16:28:53) [MSC v.1942 64 bit (AMD64)]",
-        "timestamp_utc": "2026-07-25T13:14:45.558624+00:00",
-        "houdini_version": "21.0.631",
-        "ui_available": true,
-        "copernicus_available": true,
-        "cop2_node_type_count": 156
-      }
-    },
-    {
-      "section": "cop_network_create",
-      "status": "pass",
-      "detail": "copnet created, cook=True",
-      "evidence": {
-        "copnet_path": "/img/_pip2930_probe_copnet",
-        "copnet_type": "copnet",
-        "source_type": "constant",
-        "cook_success": true,
-        "layer_resolution": "ok",
-        "has_layer_method": true,
-        "note": "gradient node fails inside copnet in H21; used constant instead"
-      }
-    },
-    {
-      "section": "opencl_cop_create_cook",
-      "status": "fail",
-      "detail": "OpenCL COP cook=False",
-      "evidence": {
-        "ocl_type": "opencl",
-        "parm_count": 38,
-        "kernel_source_set": true,
-        "kernelname_set": true,
-        "cook_success": false
-      }
-    },
-    {
-      "section": "materialx_op_path_wire",
-      "status": "pass",
-      "detail": "op: path wired to mtlximage.file",
-      "evidence": {
-        "cop_path": "/img/_pip2930_probe_copnet/_probe_output",
-        "op_reference": "op:/img/_pip2930_probe_copnet/_probe_output",
-        "file_parm_name": "file",
-        "op_ref_accepted": true,
-        "resolved_value": "op:/img/_pip2930_probe_copnet/_probe_output"
-      }
-    },
-    {
-      "section": "opencl_device_enumeration",
-      "status": "pass",
-      "detail": "hgpuinfo=True",
-      "evidence": {
-        "hgpuinfo": {
-          "available": true,
-          "returncode": 0,
-          "stdout": "[*HFS OpenCL Platform*]   Intel(R) OpenCL\nPlatform Vendor           Intel(R) Corporation\nPlatform Version          OpenCL 1.2 WINDOWS\nOpenCL Device             AMD Ryzen 9 9950X 16-Core Processor            \nOpenCL Type               CPU\nDevice Version            OpenCL 1.2 (Build 37)\nDriver Version            6.4.0.37\nFrequency                 0 MHz\nCompute Units             32\nDevice Address Bits       64\nGlobal Memory             95900 MB\nMax Allocation            23975 MB\nGlobal Cache       "
-        },
-        "hou_opencl": {
-          "available": false
-        },
-        "ocl_env_vars": {}
-      }
-    },
-    {
-      "section": "cop_output_cook_chain",
-      "status": "pass",
-      "detail": "Full COP chain evaluated",
-      "evidence": {
-        "display_set": true,
-        "layer_ok": true
-      }
-    }
-  ],
-  "summary": {
-    "pass": 5,
-    "fail": 1,
-    "blocked": 0
-  }
-}
